Refactoring

+++ js/rrl-system/pages/my-projects/my-projects-section-1.mjs
@@ -1,35 +1,60 @@
-import { RrlElement, html, css } from '../../../rrl-element.mjs'
+import { BaseElement, html, css } from '../../../base-element.mjs'
 
-class MyProjectsSection1 extends RrlElement {
+import '../../../../components/dialogs/confirm-dialog.mjs'
+import '../../../../components/input/input.mjs'
+
+class MyProjectsSection1 extends BaseElement {
         static get properties() {
             return {
                 version: { type: String, default: '1.0.0', save: true },
                 dataSet: {type: Array, default: []},
                 currentProject: {type: String, default: ""},
+                isModified: {type: Boolean, default: ""},
+                // isValidate: {type: Boolean, default: false, local: true},
             }
         }
 
         static get styles() {
             return [
-                RrlElement.styles,
+                BaseElement.styles,
                 css`
                     :host {
                         display: grid;
+                        width: 100%;
                         grid-template-columns: 3fr 9fr;
                         grid-template-rows: 50px 1fr 50px;
                         grid-template-areas:
-                            "header  header"
+                            "header1  header2"
                             "sidebar content"
                             "footer  footer";
                         gap: 0 20px;
                         background: linear-gradient(180deg, var(--header-background-color) 0%, var(--gradient-background-color) 100%);
                     }
 
-                    header {
-                        grid-area: header;
+                    header{
                         display: flex;
                         justify-content: space-between;
                         align-items: center;
+                    }
+
+                    #project-header{
+                        grid-area: header1;
+                        overflow: hidden;
+                        white-space: nowrap;
+                        text-overflow: ellipsis;
+                    }
+
+                    #project-header p {
+                        width: 100%;
+                        overflow: hidden;
+                        white-space: nowrap;
+                        text-overflow: ellipsis;
+                        font-size: 1rem;
+                        margin: 0;
+                    }
+
+                    #property-header{
+                        grid-area: header2;
                     }
 
                     .left-layout {
@@ -43,9 +68,9 @@ class MyProjectsSection1 extends RrlElement {
                         background: rgba(255, 255, 255, 0.1);
                     }
 
-
-                    .left-layout rrl-button {
+                    .left-layout simple-button {
                         width: 100%;
+                        height: 40px;
                     }
 
                     img {
@@ -55,7 +80,7 @@ class MyProjectsSection1 extends RrlElement {
                     .right-layout {
                         grid-area: content;
                         display: flex;
-                        flex-basis: 70%;
+                        justify-content: center;
                         align-items: center;
                         margin-right: 20px;
                         background: rgba(255, 255, 255, 0.1);
@@ -105,6 +130,9 @@ class MyProjectsSection1 extends RrlElement {
                         margin-right: 20px;
                         gap: 10px;
                     }
+                    footer simple-button {
+                        height: 40px;
+                    }
                 `
             ]
         }
@@ -114,35 +142,67 @@ class MyProjectsSection1 extends RrlElement {
         }
 
         async showProject(index, projectId) {
-            this.currentProject = this.dataSet[index];
+            if (this.isModified) {
+                const modalResult = await this.confirmDialogShow('Проект был изменен. Сохранить сделанные изменения?')
+                if (modalResult === 'Ok')
+                    this.saveProject().then(() => this.currentProject = this.dataSet[index]);
+            }
+            else {
+                this.currentProject = this.dataSet[index]
+            }
         }
 
         render() {
             return html`
-                <header>Прокеты: ${this.currentProject.name}</header>
+                <confirm-dialog></confirm-dialog>
+                <header id="project-header"><p>Проект: ${this.currentProject.name}</p></header>
+                <header id="property-header">Project property</header>
                 <div class="left-layout">
                     ${this.dataSet.map((project, index) =>
-                        html`<rrl-button label=${project.name} title=${project._id} height="40px" @click=${() => this.showProject(index, project._id)}></rrl-button>`)}
+                        html`<simple-button label=${project.name} title=${project._id} @click=${() => this.showProject(index, project._id)}></simple-button>`)}
                 </div>
                 <div class="right-layout">
                     <div>
-                        <h2>Innovative programming</h2>
-                        <h1>Reinforcement<br>learning<br>systems</h1>
-                        <p>The future is already here. Аrtificial intelligence never sleeps and never gets bored</p>
-                        <a href="#my-pride">Learn more</a>
+                        <h2>Project property</h2>
+                        <rrl-input id="name" type="text" name="user" @value-changed=${this.validateInput} placeholder="Название" icon="{}" fill="gray" size="20" .value=${this.currentProject.name} color="gray"></rrl-input>
+                        <rrl-input id="path" type="text" name="bars" @value-changed=${this.validateInput} placeholder="project" icon="{}" fill="gray" size="20" .value=${this.currentProject.path} color="gray"></rrl-input>
+                        <div id="drop_zone" ondrop="dropHandler(event);">
+                              <p>Drag one or more files to this <i>drop zone</i>.</p>
+                        </div>
                     </div>
                 </div>
                 <footer>
-                    <rrl-button label="Удалить" height="40px" @click=${() => this.deleteProject()}></rrl-button>
-                    <rrl-button label="Добавить" height="40px" @click=${() => this.addProject()}></rrl-button>
+                    <simple-button label=${this.isModified ? "Сохранить": "Удалить"} @click=${this.isModified ? this.saveProject: this.deleteProject}></simple-button>
+                    <simple-button label=${this.isModified ? "Отменить": "Добавить"} @click=${this.isModified ? this.cancelProject: this.addProject}></simple-button>
                 </footer>
             `;
         }
+
+        validateInput(e) {
+            if (e.target.value !== "") {
+                this.oldValues ??= new Map();
+                if (!this.oldValues.has(e.target))
+                    this.oldValues.set(e.target, this.currentProject[e.target.id])
+                else {
+                    if (this.oldValues.get(e.target) === e.target.value) {
+                        this.oldValues.delete(e.target)
+                    }
+                }
+                this.currentProject[e.target.id] = e.target.value
+                this.isModified = this.oldValues.size !== 0;
+            }
+        }
+
         async getToken() {
             return localStorage.getItem('rememberMe') ?
                 localStorage.getItem('accessUserToken') :
                 sessionStorage.getItem('accessUserToken')
         }
+
+        async confirmDialogShow(message) {
+            return await this.renderRoot.querySelector('confirm-dialog').show(message);
+        }
+
         async getProjectList() {
             const token = await this.getToken();
             return fetch('http://localhost:7000/api/projects', {
@@ -162,10 +222,14 @@ class MyProjectsSection1 extends RrlElement {
             // .then(() => this.modalDialogShow())
             .catch(err => {console.error(err.message)});
         }
+
         async saveDataSet(projects) {
+            if (projects.length === 0)
+                return;
             this.dataSet = projects.map(project =>
                 project.doc
             ).sort( (a, b) => b._id.localeCompare(a._id) )
+            this.currentProject = this.dataSet[0];
         }
 
         async addProject() {
@@ -174,7 +238,8 @@ class MyProjectsSection1 extends RrlElement {
             return fetch(`http://localhost:7000/api/project`, {
                 method: "POST",
                 headers: {
-                  'Authorization': `Bearer ${token}`
+                  'Authorization': `Bearer ${token}`,
+                  'Content-Type': 'application/json;charset=utf-8'
                 },
                 body: JSON.stringify(project)
             })
@@ -183,37 +248,94 @@ class MyProjectsSection1 extends RrlElement {
                 if (json.error) {
                     throw Error(json.error)
                 }
-                return json;
+                return json.id;
             })
-            .then(projects => console.log(projects))
+            .then(projectId => this.getProject(projectId))
+            .then(project => this.addToDataset(project))
             // .then(() => this.modalDialogShow())
             .catch(err => {console.error(err.message)});
         }
 
-        async deleteProject() {
+        async saveProject() {
             const token = await this.getToken();
-            return fetch(`http://localhost:7000/api/project/${this.currentProject._id}:${this.currentProject._rev}`, {
+            return fetch(`http://localhost:7000/api/project`, {
+                method: "PUT",
+                headers: {
+                  'Authorization': `Bearer ${token}`,
+                  'Content-Type': 'application/json;charset=utf-8'
+                },
+                body: JSON.stringify(this.currentProject)
+            })
+            .then(response => response.json())
+            .then(json => {
+                if (json.error) {
+                    throw Error(json.error)
+                }
+                return json;
+            })
+            .then(projectHeader => this.updateDataset(projectHeader))
+            // .then(() => this.modalDialogShow())
+            .catch(err => {console.error(err)});
+        }
+
+        async deleteProject() {
+            const modalResult = await this.confirmDialogShow('Вы действительно хотите удалить этот проект?')
+            if (modalResult !== 'Ok')
+                return;
+            const token = await this.getToken();
+            return fetch(`http://localhost:7000/api/project/${this.currentProject._id}?rev=${this.currentProject._rev}`, {
                 method: "DELETE",
                 headers: {
                   'Authorization': `Bearer ${token}`
                 }
-
             })
             .then(response => response.json())
             .then(json => {
                 if (json.error) {
                     throw Error(json.error)
                 }
-                return json.rows;
+                return json;
             })
-            .then(projects => console.log(projects))
-            // .then(() => this.modalDialogShow())
+            .then(project => this.deleteFromDS(project))
+            // .then(() =>c.modalDialogShowShow())
             .catch(err => {console.error(err.message)});
         }
 
-        async getProject() {
+        async cancelProject() {
+            const modalResult = await this.confirmDialogShow('Вы действительно хотите отменить все изменения?')
+            if (modalResult !== 'Ok')
+                return
+            this.oldValues.forEach( (value, key) => {
+                key.value = value;
+            });
+            this.oldValues.clear();
+            this.isModified = false;
+        }
+
+        async deleteFromDS(project) {
+            const currentIndex = this.dataSet.indexOf(this.currentProject)
+            this.currentProject = this.dataSet.length === 1 ? {} :
+                currentIndex === 0 ? this.dataSet[currentIndex + 1] : this.dataSet[currentIndex - 1]
+            this.dataSet.splice(currentIndex, 1)
+            return project
+        }
+
+        async addToDataset(project) {
+            this.dataSet.unshift(project);
+            this.currentProject = this.dataSet[0]
+            return project
+        }
+
+        async updateDataset(projectHeader) {
+            this.currentProject._id = projectHeader.id;
+            this.currentProject._rev = projectHeader.rev;
+            this.oldValues?.clear();
+            this.isModified = false;
+        }
+
+        async getProject(projectId) {
             const token = await this.getToken();
-            return fetch(`http://localhost:7000/api/project/${this.currentProject._id}}`, {
+            return fetch(`http://localhost:7000/api/project/${projectId}`, {
                 headers: {
                   'Authorization': `Bearer ${token}`
                 }
@@ -225,8 +347,11 @@ class MyProjectsSection1 extends RrlElement {
                 }
                 return json;
             })
-            .then(projects => console.log(projects))
-            // .then(() => this.modalDialogShow())
+            .then(project => {
+                console.log(project)
+                return project
+            })
+            // .then(() =>c.modalDialogShowShow())
             .catch(err => {console.error(err.message)});
         }
